Rename `HighlightedText` to `HighlightedLink`
