+++ client/web/src/fuzzy/FuzzySearch.ts
@@ -1,4 +1,4 @@
-import { HighlightedTextProps } from './HighlightedText'
+import { HighlightedLinkProps } from './HighlightedLink'
 
 export interface FuzzySearchParameters {
     query: string
@@ -8,7 +8,7 @@ export interface FuzzySearchParameters {
 }
 
 export interface FuzzySearchResult {
-    results: HighlightedTextProps[]
+    links: HighlightedLinkProps[]
     isComplete: boolean
     elapsedMilliseconds?: number
     falsePositiveRatio?: number
@@ -34,13 +34,13 @@ export interface SearchReady {
 /**
  * Superclass for different fuzzy finding algorithms.
  *
- * NOTE: Currently, there is only one implementation that uses bloom filters.
- * This implementation is specifically tailored for large repos that have >400k
- * source files. It could be that some users prefer a different fuzzy-finding
- * algorithm that is more fuzzy (relaxed with capitalization) but has worse
- * performance in XXL repos. Consider this superclass as an open invitation to
- * contribute a different fuzzy finding implementation that suits your personal
- * preferences :)
+ * Currently, there is only one implementation that is case sensitive.  This
+ * implementation is specifically tailored for large repos that have >400k
+ * source files. Most users will likely prefer case-insensitive fuzzy filtering,
+ * which is easy to support for small repos (<20k files) but it's not clear how
+ * to support that in larger repos without sacrificing latency.
+ *
+ * Tracking issue to add case-insensitive search: https://github.com/sourcegraph/sourcegraph/issues/21201
  */
 export abstract class FuzzySearch {
     public abstract totalFileCount: number
